requirements.txt
alembic==1.12.0
aniso8601==9.0.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
exceptiongroup==1.1.3
factories==1.2.0
factory-boy==3.3.0
Faker==19.4.0
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Migrate==4.0.4
Flask-RESTful==0.3.10
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.1
pluggy==1.3.0
psycopg2-binary==2.9.7
pytest==7.4.0
pytest-flask==1.2.0
pytest-timeout==2.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.20
tomli==2.0.1
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.7
zipp==3.16.2